# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    post:
      operationId: UserController_create
      summary: create new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: user successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: request body does not contain required fields
      tags: &ref_0
        - user
    get:
      operationId: UserController_findAll
      summary: get all users
      parameters: []
      responses:
        '200':
          description: return all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
      tags: *ref_0
  /user/{id}:
    get:
      operationId: UserController_findOne
      summary: get user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: return user by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: user not found
      tags: *ref_0
    put:
      operationId: UserController_update
      summary: update user password
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: user successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: user not found
        '409':
          description: old password is invalid
      tags: *ref_0
    delete:
      operationId: UserController_remove
      summary: delete user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: user successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: user not found
      tags: *ref_0
  /artist:
    post:
      operationId: ArtistController_create
      summary: create new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: artist successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: request body does not contain required fields
      tags: &ref_1
        - artist
    get:
      operationId: ArtistController_findAll
      summary: get all artists
      parameters: []
      responses:
        '200':
          description: return all artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtistResponseDto'
      tags: *ref_1
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: get artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: return artist by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: artist not found
      tags: *ref_1
    put:
      operationId: ArtistController_update
      summary: update artist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: artist successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: artist not found
      tags: *ref_1
    delete:
      operationId: ArtistController_remove
      summary: delete artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: artist successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: artist not found
      tags: *ref_1
  /track:
    post:
      operationId: TrackController_create
      summary: create new track
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: track successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: request body does not contain required fields
      tags: &ref_2
        - track
    get:
      operationId: TrackController_findAll
      summary: get all tracks
      parameters: []
      responses:
        '200':
          description: return all tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackResponseDto'
      tags: *ref_2
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: get track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: return track by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: track not found
      tags: *ref_2
    put:
      operationId: TrackController_update
      summary: update track
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: track successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: track not found
      tags: *ref_2
    delete:
      operationId: TrackController_remove
      summary: delete track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: track successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: track not found
      tags: *ref_2
  /album:
    post:
      operationId: AlbumController_create
      summary: create new album
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: album successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: request body does not contain required fields
      tags: &ref_3
        - album
    get:
      operationId: AlbumController_findAll
      summary: get all albums
      parameters: []
      responses:
        '200':
          description: return all albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumResponseDto'
      tags: *ref_3
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: get album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: return album by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: album not found
      tags: *ref_3
    put:
      operationId: AlbumController_update
      summary: update album
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: album successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: id is not UUID
        '404':
          description: album not found
      tags: *ref_3
    delete:
      operationId: AlbumController_remove
      summary: delete album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: album successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: album not found
      tags: *ref_3
  /favs/track/{id}:
    post:
      operationId: FavsController_create
      summary: add to favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: successfully added
        '400':
          description: id is not UUID
        '422':
          description: nothing to add to favs
      tags: &ref_4
        - favs
    delete:
      operationId: FavsController_remove
      summary: delete from favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: is not favorite
      tags: *ref_4
  /favs/album/{id}:
    post:
      operationId: FavsController_create
      summary: add to favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: successfully added
        '400':
          description: id is not UUID
        '422':
          description: nothing to add to favs
      tags: *ref_4
    delete:
      operationId: FavsController_remove
      summary: delete from favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: is not favorite
      tags: *ref_4
  /favs/artist/{id}:
    post:
      operationId: FavsController_create
      summary: add to favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: successfully added
        '400':
          description: id is not UUID
        '422':
          description: nothing to add to favs
      tags: *ref_4
    delete:
      operationId: FavsController_remove
      summary: delete from favs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: successfully deleted
        '400':
          description: id is not UUID
        '404':
          description: is not favorite
      tags: *ref_4
  /favs:
    get:
      operationId: FavsController_findAll
      summary: get all favs
      parameters: []
      responses:
        '200':
          description: return all favs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavsResponseDto'
      tags: *ref_4
info:
  title: Home Library Service
  description: ecivreS yrarbiL emoH
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: login123
          description: user login
        password:
          type: string
          example: password123
          description: user password
      required:
        - login
        - password
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: uuid v4
        login:
          type: string
          description: user login
        version:
          type: number
          description: version number (1..N)
        createdAt:
          type: number
          example: 1764220526750
          description: creation timestamp
        updatedAt:
          type: number
          example: 1764220526750
          description: last update timestamp
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: oldpass123
          description: current password
        newPassword:
          type: string
          example: newpass123
          description: new password
      required:
        - oldPassword
        - newPassword
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: artist name
        grammy:
          type: boolean
          description: has the artist been awarded a Grammy
      required:
        - name
        - grammy
    ArtistResponseDto:
      type: object
      properties:
        id:
          type: string
          description: uuid v4
        name:
          type: string
          description: artist name
        grammy:
          type: boolean
          description: has the artist been awarded a Grammy
      required:
        - id
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: artist name
        grammy:
          type: boolean
          description: has the artist been awarded a Grammy
      required:
        - name
        - grammy
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: track name
        artistId:
          type: string
          description: uuid v4 or null
        albumId:
          type: string
          description: uuid v4 or null
        duration:
          type: number
          description: track duration (sec)
      required:
        - name
        - artistId
        - albumId
        - duration
    TrackResponseDto:
      type: object
      properties:
        id:
          type: string
          description: uuid v4
        name:
          type: string
          description: track name
        artistId:
          type: string
          description: uuid v4 or null
        albumId:
          type: string
          description: uuid v4 or null
        duration:
          type: number
          description: track duration (sec)
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: track name
        artistId:
          type: string
          description: uuid v4 or null
        albumId:
          type: string
          description: uuid v4 or null
        duration:
          type: number
          description: track duration (sec)
      required:
        - name
        - artistId
        - albumId
        - duration
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: album name
        year:
          type: number
          description: album name
        artistId:
          type: string
          description: uuid v4 or null
      required:
        - name
        - year
        - artistId
    AlbumResponseDto:
      type: object
      properties:
        id:
          type: string
          description: uuid v4
        name:
          type: string
          description: album name
        year:
          type: number
          description: album name
        artistId:
          type: string
          description: uuid v4 or null
      required:
        - id
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: album name
        year:
          type: number
          description: album name
        artistId:
          type: string
          description: uuid v4 or null
      required:
        - name
        - year
        - artistId
    FavsResponseDto:
      type: object
      properties:
        artists:
          description: array of favorite artists
          type: array
          items:
            $ref: '#/components/schemas/ArtistResponseDto'
        albums:
          description: array of favorite albums
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponseDto'
        tracks:
          description: array of favorite tracks
          type: array
          items:
            $ref: '#/components/schemas/TrackResponseDto'
      required:
        - artists
        - albums
        - tracks